Change link implementation to use lambda functions
This change makes links simply lambda functions that are automatically generated. This makes them behave exactly the same as any other function, and they do not have to be treated as special cases, which simplifies the code and reduces cavaets.

They have a little bit of extra information that is there to help with presentation in the graph.

The `graphviz` method now takes an `expand_links` parameter which control whether the links have lables or are just small circles.

This has associated tests.